+++ src/ui/server/controllers/insurance/policy-and-export/single-contract-policy/validation/rules/index.ts
@@ -1,4 +1,4 @@
-import requestedStartDateRules from './requested-start-date';
+import requestedStartDateRules from '../../../../../../shared-validation/requested-start-date';
 import totalContractValueRules from './total-contract-value';
 import creditPeriodWithBuyerRules from './credit-period-with-buyer';
 import { ValidationErrors } from '../../../../../../../types';
